Refactor settings page into smaller chunks
