+++ website/src/components/settings/index.ts
@@ -1,0 +1,6 @@
+export { NetworkSettingsContainer } from "./network-settings-container";
+export { DeviceConsolePasswordContainer } from "./device-console-password-container";
+export { DeviceSshContainer } from "./device-ssh-container";
+export { TimeContainer } from "./time-container";
+export { LedColorContainer } from "./led-color-container";
+export { AboutContainer } from "./about-container";
